+++ people-welcome-front-end/src/components/LeftBar.jsx
