+++ people-welcome-front-end/src/components/Post.jsx
