+++ people-welcome-front-end/src/components/SearchBar.jsx
